rag-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.10
langchain-core>=0.1.8
pymilvus==2.3.3
sentence-transformers==2.5.1
transformers==4.35.2
# torch is installed via Dockerfile with CUDA support
requests==2.31.0
httpx==0.25.2
sse-starlette==1.6.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
numpy==1.24.3
PyJWT==2.8.0
asyncpg==0.29.0
aiohttp==3.9.1
Pillow==10.1.0